% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convStringToMatrix.R
\name{convStringToMatrix}
\alias{convStringToMatrix}
\title{<convStringToMatrix>}
\usage{
convStringToMatrix(encodedSeqs)
}
\arguments{
\item{encodedSeqs}{a vector of string, each string is the one-hot encoded sequence}
}
\value{
x_array_new a 3D array
}
\description{
This function converts the one-hot encoded sequence into a matrix/3D array as the input for DeepCirCode.
}
