% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepCirCode_test.R
\name{DeepCirCode_test}
\alias{DeepCirCode_test}
\title{<DeepCirCode_test>}
\usage{
DeepCirCode_test(x_test, y_test, species)
}
\arguments{
\item{x_test}{a 3D array}

\item{y_test}{a 2D array}

\item{species}{a string to be "human", or "mouse", or "fly"}
}
\description{
This function test DeepCirCode model with test case
}
